requirement.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cmake==3.18.0
Django==3.0.7
djangorestframework==3.11.0
dlib==19.20.0
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
jsonfield==3.1.0
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
